pyproject.toml
[project]
name = "cspawn"
version = "0.4.1"
description = "Manages code server containers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.4.1",
    "click>=8.1.8",
    "flask>=3.1.0",
    "flask-httpauth>=4.8.0",
    "flask-login>=0.6.3",
    "flask-pymongo>=2.3.0",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gevent>=24.11.1",
    "gunicorn>=23.0.0",
    "humanize>=4.11.0",
    "jinja2>=3.1.5",
    "jtlutil",
    #"jtlutil @ git+https://github.com/league-infrastructure/jtlutil",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy-utils>=0.41.2",
    "sqlitedict>=2.1.0",
    "tabulate>=0.9.0",
    "tenacity>=9.0.0",
    "cspawn",
    "passlib>=1.7.4",
    "flask-dance>=7.1.0",
    "flask-migrate>=4.1.0",
    "faker>=36.1.0",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["cspawn*"]  # Include all subpackages

[tool.uv.sources]
jtlutil = { path = "../jtlutil" }

[dependency-groups]
dev = [
    "ansible>=11.2.0",
    "pytest>=8.3.4",
]


[project.scripts]
cspawnctl = "cspawn.cli.ctl:cli"
